requirements.txt
PyQt5
numpy
pyrr
PyOpenGL